#5 users can now change their passwords
without using the forgot password method. added a profile menu entry in the cms, changed the link of "profile" in the main-app side menu to direct to this new profile page. the profile page only shows your own e-mail and the reset-password form. it is ugly but functional. the old reset/forgot password method works still.

noteworthy: I changed the "me" variable to "myself" in user.service.js because there is already a method with the same name; I split the reset password into resetting and mailing the link. changing the password reuses the existing reset password mechanism, it is just not sending any mail.